backend/pyproject.toml
[project]
name = "streetballer"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "pymongo>=4.10.0",
    "python-dotenv>=1.0.0",
    "pyjwt>=2.10.0",
    "argon2-cffi>=23.1.0",
    "python-crontab>=3.2.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "httpx>=0.27.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
